Some workspace config changes and housekeeping
