+++ s3db/Cargo.toml
@@ -15,7 +15,7 @@ tokio = { version = "1.36.0", default-features = false, features = ["io-util", "
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
 
-bumpalo = { version = "3.16.0", features = ["collections", "boxed"] }
+bumpalo = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = "1.4.0"
