+++ sql/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 nom = "7.1.3"
-bumpalo = {version = "3.16.0", features = ["collections", "boxed"]}
+bumpalo = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = "1.4.0"
